requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
bitsandbytes
accelerate
psutil
pyyaml

